Merge branch 'expiremental' of git://github.com/matyb/java-koans

Conflicts:
	koans-lib/src/com/sandwich/koan/constant/KoanConstants.java
	koans-lib/src/com/sandwich/koan/path/PathToEnlightenment.java
	koans-lib/src/com/sandwich/koan/path/xmltransformation/XmlToPathTransformer.java
	koans-tests/test/com/sandwich/koan/path/CommandLineTestCase.java
	koans-tests/test/com/sandwich/koan/runner/AppReadinessForDeploymentTest.java
	koans/src/PathToEnlightment.xml

+++ koans-lib/src/com/sandwich/koan/runner/RunKoans.java
@@ -100,7 +100,7 @@ public class RunKoans extends AbstractArgumentBehavior {
 					firstFailingMethod.getMethod());
 		}
 		return new KoanResultBuilder()	.level(level)
-										.numberPassing(successfull.getCount())
+										.numberPassing((int)successfull.getCount())
 										.totalNumberOfKoanMethods(getPathToEnlightement().getTotalNumberOfKoans())
 										.failingCase(firstFailingSuite)
 										.failingMethod(firstFailingMethod).message(message)
